+++ OVP/D3D9Client/VPlanet.cpp
@@ -789,7 +789,7 @@ bool vPlanet::Update (bool bMainScene)
 			// set microtexture intensity
 			double alt = cdist-size;
 			double lvl = (clouddata->microalt1-alt)/(clouddata->microalt1-clouddata->microalt0);
-			clouddata->cloudmgr->SetMicrolevel (max (0, min (1, lvl)));
+			clouddata->cloudmgr->SetMicrolevel (max (0.0, min (1.0, lvl)));
 		}
 	}
 
@@ -852,7 +852,7 @@ void vPlanet::CheckResolution()
 
 		static const double scal2 = 1.0/log(2.0);
 		const double shift = (surfmgr2 ? 6.0 : 5.0); // reduce level for tile mgr v2, because of increased patch size
-		new_patchres = min (max ((DWORD)(scal2*log(ntx)-shift),1), max_patchres);
+		new_patchres = min (max ((DWORD)(scal2*log(ntx)-shift),(DWORD)1), max_patchres);
 	}
 	if (new_patchres != patchres) {
 		if (hashaze) {
@@ -1241,13 +1241,13 @@ bool vPlanet::ModLighting (DWORD &ambient)
 	double rscale = (size-cdist)/prm.atm_href + 1.0;    // effect altitude scale (1 on ground, 0 at reference alt)
 	double amb = prm.atm_amb0 * min (1.0, (sunelev+14.0*RAD)/(20.0*RAD)); // effect magnitude (dependent on sun elevation)
 	if (amb < 0.05) return false;
-	amb = max (0, amb-0.05);
+	amb = max (0.0, amb-0.05);
 
 	DWORD addamb = (DWORD)(amb*rscale*256.0);
 	DWORD newamb = *(DWORD*)gc->GetConfigParam (CFGPRM_AMBIENTLEVEL) + addamb;
 	ambient = 0;
 	for (int i = 0; i < 4; i++)
-		ambient |= min (255, newamb) << (i<<3);
+		ambient |= min ((DWORD)255, newamb) << (i<<3);
 	return true;
 }
